+++ CHANGELOG.md
@@ -4,9 +4,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 ### Added
- - Test suite to simulate running this addon inside the context of a consuming Ember app using [ember-cli-addon-tests](https://github.com/tomdale/ember-cli-addon-tests)
- - `yarn.lock` file
- - Ember Percy for visual diff testing
+ - ...
+
+
+## [v0.8.1] - 2017-01-03
+### Added
+- Test suite to simulate running this addon inside the context of a consuming Ember app using [ember-cli-addon-tests](https://github.com/tomdale/ember-cli-addon-tests)
+- `yarn.lock` file
+- Ember Percy for visual diff testing
+
+### Updated
+ - Ember CLI v2.10.0
 
 
 ## [v0.8.0] - 2016-11-11
@@ -175,7 +183,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
  - Initial transformicons
 
 
-[unreleased]:  https://github.com/alexdiliberto/ember-cli-transformicons/compare/v0.8.0...HEAD
+[unreleased]:  https://github.com/alexdiliberto/ember-cli-transformicons/compare/v0.8.1...HEAD
+[v0.8.1]:      https://github.com/alexdiliberto/ember-cli-transformicons/compare/v0.8.0...v0.8.1
 [v0.8.0]:      https://github.com/alexdiliberto/ember-cli-transformicons/compare/v0.7.0...v0.8.0
 [v0.7.0]:      https://github.com/alexdiliberto/ember-cli-transformicons/compare/v0.6.0...v0.7.0
 [v0.6.0]:      https://github.com/alexdiliberto/ember-cli-transformicons/compare/v0.5.0...v0.6.0